requirements.txt
# production
Flask
requests

# tests
pytest
PyHamcrest
